+++ crates/selector4nix-core/src/infrastructure/config/mod.rs
@@ -6,6 +6,7 @@ mod general_raw;
 pub use credential_parsed::{AppCredential, AppCredentialEntry};
 pub use general_parsed::{
     AppConfiguration, CacheConfiguration, CacheInfoConfiguration,
-    CloudflareOptimizationConfiguration, FastlyOptimizationConfiguration, NetworkConfiguration,
-    ProxyConfiguration, ServerConfiguration, SubstituterConfiguration,
+    CloudflareCacheProxyConfiguration, CloudflareOptimizationConfiguration,
+    FastlyOptimizationConfiguration, NetworkConfiguration, ProxyConfiguration, ServerConfiguration,
+    SubstituterConfiguration,
 };